Updates for SS4 compatibility
* Add namespaces
* Update docs
* Update composer configuration for PSR-4 autoloading
* Replace SS_Log with PSR-3 Logger calls
* Fix unit tests (esp. with SS_Log missing an error which is now -> stderr)
* FIX Class name for retrieving LDAP field mappings in config
* FIX issues in API upgrades, move template
* Shorten task segments
