+++ .scrutinizer.yml
@@ -6,4 +6,4 @@ checks:
     duplication: true
 
 filter:
-  paths: [code/*, tests/*]
+  paths: [src/*, tests/*]
