+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 4.0.0 (Unreleased)
+
+New module requirements:
+
+ - Minimum SilverStripe framework version: 4.0 or above
+ - Minimum PHP version: 5.5 or above
+ - Support for PHP 7
+
 ## 3.0.0 (Unreleased)
 
 Adding features to allow writing Member data back to LDAP.
