+++ _config/legacy.yml
@@ -1,0 +1,6 @@
+---
+Name: activedirectorylegacy
+---
+SilverStripe\ORM\DatabaseAdmin:
+  classname_value_remapping:
+    LDAPGroupMapping: SilverStripe\ActiveDirectory\Model\LDAPGroupMapping
