--- code/helpers/SAMLHelper.php
@@ -1,30 +1,0 @@
-<?php
-/**
- * Class SAMLHelper
- *
- * SAMLHelper acts as a simple wrapper for the OneLogin implementation, so that we can configure
- * and inject it via the config system.
- */
-class SAMLHelper extends Object
-{
-    /**
-     * @var array
-     */
-    public static $dependencies = [
-        'SAMLConfService' => '%$SAMLConfService',
-    ];
-
-    /**
-     * @var SAMLConfService
-     */
-    public $SAMLConfService;
-
-    /**
-     * @return OneLogin_Saml2_Auth
-     */
-    public function getSAMLauth()
-    {
-        $samlConfig = $this->SAMLConfService->asArray();
-        return new \OneLogin_Saml2_Auth($samlConfig);
-    }
-}
