+++ composer.json
@@ -15,9 +15,9 @@
         }
     ],
     "require": {
-        "php": ">=5.4",
-        "silverstripe/framework": "~3.1",
-        "silverstripe/queuedjobs": "~2.6@stable",
+        "php": ">=5.5",
+        "silverstripe/framework": "^4.0@dev",
+        "silverstripe/queuedjobs": "^3.0@dev",
         "zendframework/zend-ldap": "2.5.1",
         "zendframework/zend-authentication": "2.5.1",
         "zendframework/zend-session": "2.5.1",
@@ -29,6 +29,11 @@
     "extra": {
         "branch-alias": {
             "dev-master": "4.x-dev"
+        }
+    },
+    "autoload": {
+        "psr-4": {
+            "SilverStripe\\ActiveDirectory\\": "src/"
         }
     }
 }
